Refactor utils/test
