--- utils/test/grpc.go
@@ -1,528 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package test
-
-import (
-	"context"
-	"crypto/tls"
-	"encoding/json"
-	"fmt"
-	"log"
-	"maps"
-	"os"
-	"os/exec"
-	"path"
-	"path/filepath"
-	"runtime"
-	"slices"
-	"strconv"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/cockroachdb/errors"
-	"github.com/hyperledger/fabric-x-common/tools/cryptogen"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"github.com/tedsuo/ifrit"
-	"github.com/tedsuo/ifrit/ginkgomon"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	healthgrpc "google.golang.org/grpc/health/grpc_health_v1"
-
-	"github.com/hyperledger/fabric-x-committer/utils/channel"
-	"github.com/hyperledger/fabric-x-committer/utils/connection"
-	"github.com/hyperledger/fabric-x-committer/utils/grpcservice"
-	"github.com/hyperledger/fabric-x-committer/utils/retry"
-)
-
-var (
-	// InsecureTLSConfig defines an empty tls config.
-	InsecureTLSConfig connection.TLSConfig
-	// defaultGrpcRetryProfile defines the retry policy for a gRPC client connection.
-	defaultGrpcRetryProfile retry.Profile
-
-	// OrgRootCA is the path to organization 0's TLS client credentials in the crypto materials directory.
-	OrgRootCA = filepath.Join(cryptogen.PeerOrganizationsDir, "peer-org-0",
-		cryptogen.MSPDir, cryptogen.TLSCaCertsDir, "tlspeer-org-0-CA-cert.pem")
-
-	// OrdererRootCATLSPath is the path to organization 0's orderer TLS credentials in the crypto materials directory.
-	OrdererRootCATLSPath = filepath.Join(cryptogen.OrdererOrganizationsDir,
-		"orderer-org-0", cryptogen.MSPDir, cryptogen.TLSCaCertsDir, "tlsorderer-org-0-CA-cert.pem")
-)
-
-type (
-	// GrpcServers holds the server instances and their respective configurations.
-	GrpcServers struct {
-		Servers []*grpc.Server
-		Configs []*connection.ServerConfig
-	}
-
-	// StartServerParameters defines the parameters for starting servers.
-	StartServerParameters struct {
-		TLSConfig  connection.TLSConfig
-		NumService int
-	}
-)
-
-// ServerToMultiClientConfig is used to create a multi client configuration from existing server(s)
-// given a client TLS configuration.
-func ServerToMultiClientConfig(
-	clientTLS connection.TLSConfig, servers ...*connection.ServerConfig,
-) *connection.MultiClientConfig {
-	endpoints := make([]*connection.Endpoint, len(servers))
-	for i, server := range servers {
-		endpoints[i] = &server.Endpoint
-	}
-	return &connection.MultiClientConfig{
-		TLS:       clientTLS,
-		Endpoints: endpoints,
-	}
-}
-
-// RunGrpcServerForTest starts a GRPC server using a register method.
-// It handles the cleanup of the GRPC server at the end of a test, and ensure the test is ended
-// only when the GRPC server is down.
-// It also updates the server config endpoint port to the actual port if the configuration
-// did not specify a port.
-// The method asserts that the GRPC server did not end with failure.
-func RunGrpcServerForTest(
-	ctx context.Context, tb testing.TB, serverConfig *connection.ServerConfig, register func(server *grpc.Server),
-) *grpc.Server {
-	tb.Helper()
-	listener, err := serverConfig.Listener(ctx)
-	require.NoError(tb, err)
-	server, err := serverConfig.GrpcServer(nil)
-	require.NoError(tb, err)
-
-	if register != nil {
-		register(server)
-	} else {
-		healthgrpc.RegisterHealthServer(server, connection.DefaultHealthCheckService())
-	}
-
-	var wg sync.WaitGroup
-	tb.Cleanup(wg.Wait)
-	tb.Cleanup(server.Stop)
-
-	// The parent error capture the caller stack trace,
-	// which helps track the server origin when debugging test failures.
-	parentErr := errors.New("parent stack context")
-	wg.Go(func() {
-		// We use assert to prevent panicking for cleanup errors.
-		serveErr := server.Serve(listener)
-		if serveErr != nil && !errors.Is(serveErr, grpc.ErrServerStopped) {
-			assert.NoError(tb, errors.WithSecondaryError(serveErr, parentErr))
-		}
-	})
-
-	_ = context.AfterFunc(ctx, func() {
-		server.Stop()
-	})
-	return server
-}
-
-// StartGrpcServersForTest starts multiple GRPC servers with a default configuration.
-func StartGrpcServersForTest(
-	ctx context.Context,
-	t *testing.T,
-	p StartServerParameters,
-	register func(*grpc.Server),
-) *GrpcServers {
-	t.Helper()
-	sc := make([]*connection.ServerConfig, p.NumService)
-	for i := range sc {
-		sc[i] = NewLocalHostServer(p.TLSConfig)
-	}
-	return StartGrpcServersWithConfigForTest(ctx, t, register, sc...)
-}
-
-// StartGrpcServersWithConfigForTest starts multiple GRPC servers with given configurations.
-func StartGrpcServersWithConfigForTest(
-	ctx context.Context, t *testing.T, register func(*grpc.Server), sc ...*connection.ServerConfig,
-) *GrpcServers {
-	t.Helper()
-	grpcServers := make([]*grpc.Server, len(sc))
-	if register == nil {
-		register = func(server *grpc.Server) {
-			healthgrpc.RegisterHealthServer(server, connection.DefaultHealthCheckService())
-		}
-	}
-	for i, s := range sc {
-		grpcServers[i] = RunGrpcServerForTest(ctx, t, s, register)
-	}
-	return &GrpcServers{
-		Servers: grpcServers,
-		Configs: sc,
-	}
-}
-
-// RunServiceForTest runs a service using the given service method, and waits for it to be ready
-// given the waitFunc method.
-// It handles the cleanup of the service at the end of a test, and ensure the test is ended
-// only when the service return.
-// The method asserts that the service did not end with failure.
-// Returns a ready flag that indicate that the service is done.
-func RunServiceForTest(
-	ctx context.Context,
-	tb testing.TB,
-	service func(ctx context.Context) error,
-	waitFunc func(ctx context.Context) bool,
-) *channel.Ready {
-	tb.Helper()
-	doneFlag := channel.NewReady()
-	var wg sync.WaitGroup
-	// NOTE: we should cancel the context before waiting for the completion. Therefore, the
-	//       order of cleanup matters, which is last added first called.
-	tb.Cleanup(wg.Wait)
-	dCtx, cancel := context.WithCancel(ctx)
-	tb.Cleanup(cancel)
-
-	// We extract caller information to ensure we have sufficient information for debugging.
-	pc, file, no, ok := runtime.Caller(1)
-	require.True(tb, ok)
-	wg.Go(func() {
-		defer doneFlag.SignalReady()
-		// We use assert to prevent panicking for cleanup errors.
-		assert.NoErrorf(tb, service(dCtx), "called from %s:%d\n\t%s", file, no, runtime.FuncForPC(pc).Name())
-	})
-
-	if waitFunc == nil {
-		return doneFlag
-	}
-
-	initCtx, initCancel := context.WithTimeout(dCtx, 2*time.Minute)
-	tb.Cleanup(initCancel)
-	require.True(tb, waitFunc(initCtx), "service is not ready")
-	return doneFlag
-}
-
-// RunServiceAndGrpcForTest combines running a service and its GRPC server.
-// It is intended for services that implements the Service API (i.e., command line services).
-func RunServiceAndGrpcForTest(
-	ctx context.Context,
-	t *testing.T,
-	service grpcservice.Service,
-	serverConfig ...*connection.ServerConfig,
-) *channel.Ready {
-	t.Helper()
-	doneFlag := RunServiceForTest(ctx, t, func(ctx context.Context) error {
-		return connection.FilterStreamRPCError(service.Run(ctx))
-	}, service.WaitForReady)
-	for _, server := range serverConfig {
-		RunGrpcServerForTest(ctx, t, server, service.RegisterService)
-	}
-	return doneFlag
-}
-
-// CheckServerStopped returns true if the grpc server listening on a
-// given address has been stopped.
-func CheckServerStopped(t *testing.T, addr string) bool {
-	t.Helper()
-	ctx, cancel := context.WithTimeout(t.Context(), 2*time.Second)
-	defer cancel()
-
-	conn, err := grpc.DialContext( //nolint:staticcheck
-		ctx,
-		addr,
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithBlock(), //nolint:staticcheck
-	)
-	if err != nil {
-		return true
-	}
-	_ = conn.Close()
-	return false
-}
-
-// NewSecuredConnection creates the default connection with given transport credentials.
-func NewSecuredConnection(
-	t *testing.T,
-	endpoint connection.WithAddress,
-	tlsConfig connection.TLSConfig,
-) *grpc.ClientConn {
-	t.Helper()
-	return NewSecuredConnectionWithRetry(t, endpoint, tlsConfig, defaultGrpcRetryProfile)
-}
-
-// NewSecuredConnectionWithRetry creates the default connection with given transport credentials.
-func NewSecuredConnectionWithRetry(
-	t *testing.T,
-	endpoint connection.WithAddress,
-	tlsConfig connection.TLSConfig,
-	retryProfile retry.Profile,
-) *grpc.ClientConn {
-	t.Helper()
-	clientCreds, err := tlsConfig.ClientCredentials()
-	require.NoError(t, err)
-	conn, err := connection.NewConnection(connection.ClientParameters{
-		Address: endpoint.Address(),
-		Creds:   clientCreds,
-		Retry:   &retryProfile,
-	})
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		_ = conn.Close()
-	})
-	return conn
-}
-
-// NewInsecureConnection creates the default connection with insecure credentials.
-func NewInsecureConnection(t *testing.T, endpoint connection.WithAddress) *grpc.ClientConn {
-	t.Helper()
-	return NewInsecureConnectionWithRetry(t, endpoint, defaultGrpcRetryProfile)
-}
-
-// NewInsecureConnectionWithRetry creates the default dial config with insecure credentials.
-func NewInsecureConnectionWithRetry(
-	t *testing.T, endpoint connection.WithAddress, retryProfile retry.Profile,
-) *grpc.ClientConn {
-	t.Helper()
-	conn, err := connection.NewConnection(connection.ClientParameters{
-		Address: endpoint.Address(),
-		Creds:   insecure.NewCredentials(),
-		Retry:   &retryProfile,
-	})
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		_ = conn.Close()
-	})
-	return conn
-}
-
-// NewInsecureLoadBalancedConnection creates the default connection with insecure credentials.
-func NewInsecureLoadBalancedConnection(t *testing.T, endpoints []*connection.Endpoint) *grpc.ClientConn {
-	t.Helper()
-	conn, err := connection.NewLoadBalancedConnection(&connection.MultiClientConfig{
-		Endpoints: endpoints,
-		Retry:     &defaultGrpcRetryProfile,
-	})
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		_ = conn.Close()
-	})
-	return conn
-}
-
-// NewTLSMultiClientConfig creates a multi client configuration for test purposes
-// given number of endpoints and a TLS configuration.
-func NewTLSMultiClientConfig(
-	tlsConfig connection.TLSConfig,
-	ep ...*connection.Endpoint,
-) *connection.MultiClientConfig {
-	return &connection.MultiClientConfig{
-		Endpoints: ep,
-		TLS:       tlsConfig,
-		Retry:     &defaultGrpcRetryProfile,
-	}
-}
-
-// NewInsecureClientConfig creates a client configuration for test purposes given an endpoint.
-func NewInsecureClientConfig(ep *connection.Endpoint) *connection.ClientConfig {
-	return NewTLSClientConfig(InsecureTLSConfig, ep)
-}
-
-// NewTLSClientConfig creates a client configuration for test purposes given a single endpoint and creds.
-func NewTLSClientConfig(tlsConfig connection.TLSConfig, ep *connection.Endpoint) *connection.ClientConfig {
-	return &connection.ClientConfig{
-		Endpoint: ep,
-		TLS:      tlsConfig,
-		Retry:    &defaultGrpcRetryProfile,
-	}
-}
-
-// LogStruct logs a struct in a flat representation.
-func LogStruct(t *testing.T, name string, v any) {
-	t.Helper()
-	// Marshal struct to JSON
-	data, err := json.Marshal(v)
-	require.NoError(t, err)
-
-	// Unmarshal to map
-	var m map[string]any
-	err = json.Unmarshal(data, &m)
-	require.NoError(t, err)
-
-	// Flatten
-	flat := make(map[string]any)
-	flatten("", m, flat)
-	sb := &strings.Builder{}
-	for _, k := range slices.Sorted(maps.Keys(flat)) {
-		sb.WriteString(k)
-		sb.WriteString(": ")
-		_, printErr := fmt.Fprintf(sb, "%v", flat[k])
-		require.NoError(t, printErr)
-		sb.WriteString("\n")
-	}
-	t.Logf("%s:\n%s", name, sb.String())
-}
-
-func flatten(prefix string, in any, out map[string]any) {
-	switch val := in.(type) {
-	default:
-		out[prefix] = val
-	case map[string]any:
-		e := flattenEndpoint(val)
-		if e != nil {
-			out[prefix] = e
-			return
-		}
-		for k, v := range val {
-			key := k
-			if prefix != "" {
-				key = prefix + "." + k
-			}
-			flatten(key, v, out)
-		}
-	case []any:
-		for i, item := range val {
-			key := fmt.Sprintf("%s.%d", prefix, i)
-			flatten(key, item, out)
-		}
-	}
-}
-
-func flattenEndpoint(in map[string]any) *connection.Endpoint {
-	if len(in) != 2 {
-		return nil
-	}
-	host, okHost := in["host"]
-	if !okHost {
-		return nil
-	}
-	hostStr, okHostStr := host.(string)
-	if !okHostStr {
-		return nil
-	}
-	port, okPort := in["port"]
-	if !okPort {
-		return nil
-	}
-	portFloat, okPortFloat := port.(float64)
-	if !okPortFloat {
-		return nil
-	}
-	return &connection.Endpoint{Host: hostStr, Port: int(portFloat)}
-}
-
-const (
-	// CreatorCertificate denotes Creator field in protoblocktx.Identity to contain x509 certificate.
-	CreatorCertificate = 0
-	// CreatorID denotes Creator field in protoblocktx.Identity to contain the digest of x509 certificate.
-	CreatorID = 1
-)
-
-// MustGetTLSConfig creates a tls.Config from a connection.TLSConfig while ensuring no error return from that process.
-func MustGetTLSConfig(t *testing.T, tlsConfig *connection.TLSConfig) *tls.Config {
-	t.Helper()
-	if tlsConfig == nil {
-		return nil
-	}
-	tlsCreds, err := connection.NewClientTLSCredentials(*tlsConfig)
-	require.NoError(t, err)
-	clientTLSConfig, err := tlsCreds.CreateClientTLSConfig()
-	require.NoError(t, err)
-	return clientTLSConfig
-}
-
-// NewPreAllocatedLocalHostServer create a localhost server config with a pre allocated listener and port.
-func NewPreAllocatedLocalHostServer(t *testing.T, tlsConfig connection.TLSConfig) *connection.ServerConfig {
-	t.Helper()
-	server := NewLocalHostServer(tlsConfig)
-	listener, err := server.PreAllocateListener(t.Context())
-	t.Cleanup(func() {
-		_ = listener.Close()
-	})
-	require.NoError(t, err)
-	return server
-}
-
-// Run executes the specified command and returns the corresponding process.
-// It is important to note that the underlying invocation function (Invoke)
-// returns only when either process.Ready or process.Wait has been read.
-// Consequently, the caller only needs to read process.Wait to wait for the
-// process to complete and capture any errors that may have occurred during execution.
-//
-//nolint:ireturn
-func Run(cmd *exec.Cmd, name, startCheck string) ifrit.Process {
-	return ifrit.Invoke(ginkgomon.New(ginkgomon.Config{
-		Command:           cmd,
-		Name:              name,
-		AnsiColorCode:     "",
-		StartCheck:        startCheck,
-		StartCheckTimeout: 0,
-		Cleanup: func() {
-		},
-	}))
-}
-
-// Make runs a make command with the specified rules.
-func Make(rules ...string) error {
-	dir, err := os.Getwd()
-	if err != nil {
-		return errors.Wrap(err, "failed to get current working directory")
-	}
-	buildCmd := exec.Command("make", rules...)
-	buildCmd.Dir = path.Clean(path.Join(dir, "../.."))
-	makeRun := Run(buildCmd, "make", "")
-	log.Println("wait")
-	select {
-	case err = <-makeRun.Wait():
-		return errors.Wrap(err, "failed to run make command")
-	case <-time.After(3 * time.Minute):
-		makeRun.Signal(os.Kill)
-		return errors.New("make command timed out")
-	}
-}
-
-// NewServiceTLSConfig creates a server TLS configuration with certificates loaded from the artifact path.
-// This function constructs paths to TLS certificates for a given service within the peer organization structure.
-func NewServiceTLSConfig(artifactsPath, serviceName, mode string) connection.TLSConfig {
-	subPath := filepath.Join(artifactsPath, cryptogen.PeerOrganizationsDir, "peer-org-0",
-		cryptogen.PeerNodesDir, serviceName, cryptogen.TLSDir)
-	return connection.TLSConfig{
-		Mode:     mode,
-		CertPath: filepath.Join(subPath, "server.crt"),
-		KeyPath:  filepath.Join(subPath, "server.key"),
-		CACertPaths: []string{
-			filepath.Join(artifactsPath, OrgRootCA),
-		},
-	}
-}
-
-// OrgClientTLSConfig creates a mutual TLS client configuration using a specific
-// peer organization's TLS client certificate. The serverCACertPaths are the CA
-// certs needed to verify the server (typically from the CredentialsFactory).
-func OrgClientTLSConfig(artifactsPath string, orgIndex int, serverCACertPaths []string) connection.TLSConfig {
-	orgName := fmt.Sprintf("peer-org-%d", orgIndex)
-	orgDomain := fmt.Sprintf("peer-org-%d.com", orgIndex)
-	tlsDir := filepath.Join(artifactsPath, cryptogen.PeerOrganizationsDir, orgName,
-		cryptogen.UsersDir, fmt.Sprintf("client@%s", orgDomain), cryptogen.TLSDir)
-	return connection.TLSConfig{
-		Mode:        connection.MutualTLSMode,
-		CertPath:    filepath.Join(tlsDir, "client.crt"),
-		KeyPath:     filepath.Join(tlsDir, "client.key"),
-		CACertPaths: serverCACertPaths,
-	}
-}
-
-// NewEndpoint creates an endpoint from give host and port (as string).
-func NewEndpoint(t *testing.T, host, port string) *connection.Endpoint {
-	t.Helper()
-	convertedPort, err := strconv.Atoi(port)
-	require.NoError(t, err, "could not convert port to integer")
-	return &connection.Endpoint{Host: host, Port: convertedPort}
-}
-
-// NewLocalHostServer returns a default server config with endpoint "localhost:0" given server credentials.
-func NewLocalHostServer(creds connection.TLSConfig) *connection.ServerConfig {
-	return &connection.ServerConfig{
-		Endpoint: connection.Endpoint{Host: "127.0.0.1"},
-		TLS:      creds,
-	}
-}
